Added Spring Security in Journal

+++ src/main/java/com/example/FirstSpringApp/service/JournalEntryService.java
@@ -28,7 +28,7 @@ public class JournalEntryService {
           journalEntry.setDate(LocalDateTime.now());
           JournalEntry saved = journalEntryRepository.save(journalEntry);
           user.getJournalEntries().add(saved);
-          userService.saveEntry(user);
+          userService.saveNewUser(user);
       } catch (Exception e) {
           throw new RuntimeException(e);
       }
@@ -44,11 +44,27 @@ public class JournalEntryService {
     {
         return journalEntryRepository.findById(id);
     }
-    public void deleteById(ObjectId id, String userName)
+
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName)
     {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+
+            if(removed)
+            {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error Occured while deleting the entry.",e);
+        }
+            return removed;
     }
+
+    public List<JournalEntry> findByUserName(String userName)
+    {};
 }
